backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
boto3==1.34.0
python-dotenv==1.0.0
pyspark==3.5.1
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
imbalanced-learn==0.11.0
numpy==1.26.2
